SpyAgent 2.11.0
Added the ability to enter commands more conveniently by pressing the f12 button (you can change it in the config). when you click, simpledialog.askstring will be opened, which will work as a standard command entry in the file, but if you make a typo, it will not send the text to the guild.

The unnecessary get_history function has been removed and replaced with the use of a function from the command module.

now it is possible to understand the message more easily. now all additional information about the message, such as attached files, will be written on the next line with the "↳" symbol at the beginning.

Now help is in the team manager, and does not work separately from it.

minor changes in the local team manager.
Now the function is executed immediately instead of returning it.

Now writing the channel id can be configured in the config.